tenali-llm/requirements.txt

# Core ML Libraries
torch>=2.1.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.15.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Training & Optimization
deepspeed>=0.12.0
flash-attn>=2.3.0
xformers>=0.0.22

# Data Collection
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Economic Data
pandas-datareader>=0.10.0
fredapi>=0.5.2
alpha-vantage>=2.3.1

# Crypto Data
ccxt>=4.5.20
web3>=6.11.0

# Technical Analysis
# ta-lib  <-- Removed: Replaced with pure Pandas implementation for easier installation
pandas-ta>=0.3.14b0

# API & Deployment
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Utilities
tqdm>=4.66.0
wandb>=0.16.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
